refactor (dbserver/handlers): breakdown large files in handlers into smaller ones
