+++ internal/dbserver/handlers/health.go
@@ -1,0 +1,56 @@
+package handlers
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/gin-gonic/gin"
+)
+
+// HealthCheck provides a health check endpoint for the database server
+func (h *Handler) HealthCheck(c *gin.Context) {
+	startTime := time.Now()
+
+	// Check database connection by executing a simple query
+	dbStatus := "healthy"
+	dbError := ""
+
+	// Use a simple system query to test the connection
+	query := h.db.Session().Query("SELECT now() FROM system.local")
+	var timestamp time.Time
+	if err := query.Scan(&timestamp); err != nil {
+		dbStatus = "unhealthy"
+		dbError = err.Error()
+		h.logger.Errorf("Database health check failed: %v", err)
+	}
+
+	// Prepare response
+	response := gin.H{
+		"status":    "ok",
+		"timestamp": startTime.Unix(),
+		"service":   "dbserver",
+		"version":   "1.0.0",
+		"uptime":    time.Since(startTime).String(),
+		"database": gin.H{
+			"status": dbStatus,
+			"error":  dbError,
+		},
+		"checks": gin.H{
+			"database_connection": dbStatus == "healthy",
+		},
+	}
+
+	// Set appropriate HTTP status
+	httpStatus := http.StatusOK
+	if dbStatus != "healthy" {
+		httpStatus = http.StatusServiceUnavailable
+		response["status"] = "degraded"
+	}
+
+	// Log health check
+	duration := time.Since(startTime)
+	h.logger.Infof("Health check completed: status=%s, db_status=%s, duration=%v",
+		response["status"], dbStatus, duration)
+
+	c.JSON(httpStatus, response)
+}
