+++ internal/dbserver/handlers/keeper_update.go
@@ -1,0 +1,96 @@
+package handlers
+
+import (
+	"net/http"
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+	"github.com/trigg3rX/triggerx-backend/internal/dbserver/types"
+)
+
+func (h *Handler) IncrementKeeperTaskCount(c *gin.Context) {
+	keeperID := c.Param("id")
+	h.logger.Infof("[IncrementKeeperTaskCount] Incrementing task count for keeper with ID: %s", keeperID)
+
+	keeperIDInt, err := strconv.ParseInt(keeperID, 10, 64)
+	if err != nil {
+		h.logger.Errorf("[IncrementKeeperTaskCount] Error parsing keeper ID: %v", err)
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	newCount, err := h.keeperRepository.IncrementKeeperTaskCount(keeperIDInt)
+	if err != nil {
+		h.logger.Errorf("[IncrementKeeperTaskCount] Error retrieving current task count: %v", err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	h.logger.Infof("[IncrementKeeperTaskCount] Successfully incremented task count to %d for keeper ID: %s", newCount, keeperID)
+	c.JSON(http.StatusOK, gin.H{"no_executed_tasks": newCount})
+}
+
+func (h *Handler) AddTaskFeeToKeeperPoints(c *gin.Context) {
+	keeperID := c.Param("id")
+	keeperIDInt, err := strconv.ParseInt(keeperID, 10, 64)
+	if err != nil {
+		h.logger.Errorf("[AddTaskFeeToKeeperPoints] Error parsing keeper ID: %v", err)
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	var requestBody struct {
+		TaskID int64 `json:"task_id"`
+	}
+
+	if err := c.ShouldBindJSON(&requestBody); err != nil {
+		h.logger.Errorf("[AddTaskFeeToKeeperPoints] Error decoding request body: %v", err)
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	taskID := requestBody.TaskID
+	h.logger.Infof("[AddTaskFeeToKeeperPoints] Processing task fee for task ID %d to keeper with ID: %s", taskID, keeperID)
+
+	taskFee, err := h.taskRepository.GetTaskFee(taskID)
+	if err != nil {
+		h.logger.Errorf("[AddTaskFeeToKeeperPoints] Error retrieving task fee for task ID %d: %v", taskID, err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	newPoints, err := h.keeperRepository.UpdateKeeperPoints(keeperIDInt, taskFee)
+	if err != nil {
+		h.logger.Errorf("[AddTaskFeeToKeeperPoints] Error retrieving current points: %v", err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+
+	h.logger.Infof("[AddTaskFeeToKeeperPoints] Successfully added task fee %d from task ID %d to keeper ID: %s, new points: %d",
+		taskFee, taskID, keeperID, newPoints)
+	c.JSON(http.StatusOK, gin.H{
+		"task_id":       taskID,
+		"task_fee":      taskFee,
+		"keeper_points": newPoints,
+	})
+}
+
+func (h *Handler) UpdateKeeperChatID(c *gin.Context) {
+	var requestData types.UpdateKeeperChatIDRequest
+	if err := c.ShouldBindJSON(&requestData); err != nil {
+		h.logger.Errorf("[UpdateKeeperChatID] Error decoding request body: %v", err)
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	err := h.keeperRepository.UpdateKeeperChatID(requestData.KeeperAddress, requestData.ChatID)
+	if err != nil {
+		h.logger.Errorf("[UpdateKeeperChatID] Error updating chat ID for keeper: %s", requestData.KeeperAddress)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	h.logger.Infof("[UpdateKeeperChatID] Successfully updated chat ID for keeper: %s", requestData.KeeperAddress)
+	c.JSON(http.StatusOK, gin.H{"message": "Chat ID updated successfully"})
+}
